Add a program irregular_verbs.py that allows to study and practice english irregular verbs. Add a file irregular_verbs that contains english irregular verbs.
The program reads a file and stores all lines(v1, v2,v3 form of verb) in a list by using a function.
Then the random verb is generated by a function and it displays on the screen(v1). The user should write
V2 and V3 forms that compare with original verb. If all is correct, the verb removes from the list and the
new verb is generated by a function. Otherwise, the user can takes a hint or writes a verb again or skips it.